Adding sleep-before-ready config, and better failback for hostname.
In some environments, a race condition can exist between when a vRA
instance is provisioned to when it is actually ready. Such a race
condition may be DNS propagation. In the event of a NXDOMAIN when
TK looks up a hostname, it will cause TK to throw an exception
rather than waiting and trying again (as TK assumes a DNS failure
is unrecoverable). The `server_ready_sleep_time` allows users to
configure a sleep time before TK checks to see if the host is ready.
This is a hack, but it's an easy one.

Additionally, when `use_dns` is false, the vRA driver will still
fallback to the hostname in the event an IP address is not found.
This avoids the user needing to set use_dns to true in their
environment in these cases.